build: Support compiling without the legacy v0 compatibility code
