+++ src/client.c
@@ -2,7 +2,6 @@
 #include "assert.h"
 #include "configuration.h"
 #include "err.h"
-#include "legacy.h"
 #include "log.h"
 #include "membership.h"
 #include "progress.h"
